FIX: Cleaning up use of printf() format specifiers. %i mostly replaced by %u when appropriate. FIX: 'size_t to %i' printf() vulnerability. FIX: Subdivision, change line&point handling, add missing parentheses to clarify order. Add missing headers to vc9 workspace.
git-svn-id: https://assimp.svn.sourceforge.net/svnroot/assimp/trunk@533 67173fc5-114c-0410-ac8e-9d2fd5bffc1f

+++ code/ACLoader.cpp
@@ -176,7 +176,8 @@ void AC3DImporter::LoadObjectSection(std::vector<Object>& objects)
 		light->mAttenuationConstant = 1.f;
 
 		// Generate a default name for both the light source and the node
-		light->mName.length = ::sprintf(light->mName.data,"ACLight_%i",mLights->size()-1);
+		// FIXME - what's the right way to print a size_t? Is 'zu' universally available? stick with the safe version.
+		light->mName.length = ::sprintf(light->mName.data,"ACLight_%i",static_cast<unsigned int>(mLights->size())-1);
 		obj.name = std::string( light->mName.data );
 
 		DefaultLogger::get()->debug("AC3D: Light source encountered");
